Oops, fix Shop Wizard URL!
I'm not sure when TNT changed this, but the Shop Wizard has a new
canonical URL now! The previous one redirects, but it does *not* pass
along query string parameters, so these buttons would correctly go to
the Shop Wizard but not auto-fill the name!

Now, we go directly to the new canonical URL, with query params in hand!